Moved to chunked encoding.
Refactor mgos_prometheus_metrics_printf() to output a chunk, rather than
the string itself. An HTTP chunk is written as a sequence of:
  mg_printf(nc, "%X\r\n%s\r\n", chunklen, chunk);

As a compromise, the printer now needs to format the string in order to
determine its length, which does cause some memory bloat, and will
obviously truncate metrics which exceed 500 bytes in length.

Add metrics_send_chunks() which sends the system chunks and loops over
the registered callbacks, ultimately sending the empty chunk (0\r\n\r\n) to
the mg_connection.

This should be non-intrusive for users, who will keep on using
mgos_prometheus_metrics_printf().

This is in preparation of adding prometheus.pushgateway functionality in
an upcoming commit. We need chunked encoding because the HTTP POST
request does not have a known content length up front, and HTTP requires
us to either specify Content-Length, or enable Content-Encoding:
chunked.