Add name cache
Prometheus parser will only allow a metric to have one TYPE and HELP line,
which means that if applicatinos call mgos_prometheus_metrics_printf()
multiple times on the same name, the parser borks. Therefore, we have to
keep a cache of the names we've already output. This is implemented in
cache.[ch]

Add unittests for the cache implementation.